Added basic render functionality
I added the renderer class. It contains an array of Colors called colorMap, as well as a function that takes a 2D vector of indices and returns a color, and a function that generates a random 2D vector of ints for testing. I added code in main that displays the image rendered.